Update test files
