+++ internal/services/compute/orchestrated_virtual_machine_scale_set_disk_data_resource_test.go
