+++ internal/services/compute/orchestrated_virtual_machine_scale_set_extensions_resource_test.go
