+++ internal/services/compute/orchestrated_virtual_machine_scale_set_identity_resource_test.go
