+++ internal/services/compute/orchestrated_virtual_machine_scale_set_legacy_resource_test.go
