+++ internal/services/compute/orchestrated_virtual_machine_scale_set_other_resource_test.go
